Search normalized paths for parameters as well
Parameters can be defined for all verbs within a path at the path level
in the OpenAPI spec. E.g.

```
paths:
  /user/{userId}:
    parameters:
      ....

    get: # parameters apply here as well
```

However, the typegen code wouldn't find those because it was only
searching for exported types under the operationId and didn't include
searches that apply to the overall path. This change adds directives to
check the path-level for applicable request parameters as well.